+++ test/system/underline_panels_test.rb
@@ -2,8 +2,8 @@
 
 require "application_system_test_case"
 
-class IntegrationUnderlinePanelTest < ApplicationSystemTestCase
-  def assert_underline_panel_rendered
+class IntegrationUnderlinePanelsTest < ApplicationSystemTestCase
+  def assert_underline_panels_rendered
     assert_selector("tab-container") do
       assert_selector("div.UnderlineNav") do
         assert_selector("ul.UnderlineNav-body[role='tablist']") do
@@ -43,13 +43,13 @@ class IntegrationUnderlinePanelTest < ApplicationSystemTestCase
   def test_renders_component
     with_preview(:default)
 
-    assert_underline_panel_rendered
+    assert_underline_panels_rendered
   end
 
   def test_changes_tabs_on_click
     with_preview(:default)
 
-    assert_underline_panel_rendered
+    assert_underline_panels_rendered
 
     assert_selects_tab(2)
     assert_selects_tab(3)